Follow up fixes after PR #797
Test: existing tests
