+++ testProjectAndroidBase/build_base.gradle
@@ -117,7 +117,7 @@ dependencies {
 def assertJavaCompileHasProtoGeneratedDir(Object variant, Collection<String> codegenPlugins) {
   String taskName = "generate${variant.name.capitalize()}Proto"
   def task = project.tasks.getByName(taskName)
-  def baseDir = task.outputBaseDir.get().asFile
+  def baseDir = task.outputBaseDirProperty.get().asFile
   rootProject.assertJavaCompileHasProtoGeneratedDir(baseDir, variant.javaCompileProvider.get().source , codegenPlugins)
 }
